updating readme
I finally decided to write a better readme to this project. So far, I think this is the coolest thing I did in programming, I am not sure if it is the best code, since my java game is pretty good by its own, although not as cool as this.